fix: Modularizamos actualizar_lobby en main
- Pasamos las funcionalidades de actualizar_lobby dentro del main

 Cambios a ser confirmados:
	borrados:        actualizar_lobby.py
	modificados:     main.py
	renombrados:     test_actualizar_lobby.py -> tests/test_actualizar_lobby.py
