+++ main.py
@@ -4,6 +4,9 @@ from sqlalchemy.orm import Session
 
 from database import Database
 from repositories import GameRepository
+from services.connection_manager import ManejadorConexionesLobby
+
+manejador = ManejadorConexionesLobby()
 
 db_uri = getenv("DB_URI")
 if db_uri is not None:
@@ -30,3 +33,20 @@ def get_games_repo(request: Request) -> GameRepository:
 def get_games_available(repo: GameRepository = Depends(get_games_repo)):
     lobbies = repo.get_available(10)
     return lobbies
+
+@app.websocket("/ws/lobby/{lobby_id}")
+async def websocket_endpoint(websocket: WebSocket, lobby_id: int):
+    # Conectar al nuevo jugador a la lista de conexiones del lobby
+    await manejador.conectar(websocket, lobby_id)
+
+    await manejador.broadcast(f"Un nuevo jugador ha ingresado al lobby {lobby_id}.", lobby_id)
+    
+    try:
+        while True:
+            # Esto solo lo ponemos para mantener la conexión abierta
+            await websocket.receive_text()
+    except WebSocketDisconnect:
+        # Desconectar al jugador y notificamos a los demás jugadores
+        manejador.desconectar(websocket, lobby_id)
+
+        await manejador.broadcast(f"Un jugador ha abandonado el lobby {lobby_id}.", lobby_id)
